Added create byand delete in post back & front end

+++ backend/routes/api/posts.js
@@ -34,7 +34,8 @@ router.post("/", checkAuth, multer({storage: storage}).single('image'), (req, re
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   //console.log(post);
   post.save().then(createdPost => {
@@ -61,9 +62,12 @@ router.put("/:id", checkAuth, multer({storage: storage}).single('image') ,(req, 
     content: req.body.content,
     imagePath: imagePath
   });
-  console.log(post);
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    res.status(200).json({ message: 'Update Successful!' });
+  Post.updateOne({ _id: req.params.id , creator: req.userData.userId}, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({ message: 'Update Successful!' });
+    } else {
+      res.status(401).json({ message: 'Not Authorized!' });
+    }
   })
 });
 
@@ -100,8 +104,12 @@ router.get('/:id', checkAuth, (req, res, next) => {
 
 router.delete('/:id', checkAuth, (req, res, next) => {
   console.log(req.params.id);
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    res.status(200).json({ message: 'post deleted!' })
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({ message: 'Deletion Successful!' });
+    } else {
+      res.status(401).json({ message: 'Not Authorized!' });
+    }
   });
 });
